Move patch functionality into patch*.go
